+++ src/banco/modelo/pagamento/Boleto.java
@@ -7,16 +7,16 @@ import banco.modelo.Pessoa;
 public class Boleto implements DocumentoPagavel, DocumentoEstornavel{
 	
     private Pessoa beneficiario;
-    private double valor;
+    private BigDecimal valor;
     private boolean pago;
 
-    public Boleto(Pessoa beneficiario, double valor) {
+    public Boleto(Pessoa beneficiario, BigDecimal valor) {
         this.beneficiario = beneficiario;
         this.valor = valor;
     }
 
 	@Override
-	public double getValorTotal() {
+	public BigDecimal getValorTotal() {
 		return valor;
 	}
